requirements.txt
streamlit
yfinance
pandas
plotly
requests
beautifulsoup4
langchain
langchain-huggingface
langchain-core
python-dotenv


